refactor(renderer): extract App stream runtime orchestration
